requirements.txt
fastapi
jinja2
uvicorn
requests
sqlalchemy
pydantic
rich
python-jose[cryptography]
bcrypt
python-multipart
